pyproject.toml
[project]
name = "canopy-web"
version = "0.1.0"
description = "Canopy Web - Django ASGI backend with React frontend"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "django-environ>=0.11,<1.0",
    "django-cors-headers>=4.3,<5.0",
    "uvicorn[standard]>=0.27,<1.0",
    "psycopg[binary]>=3.1,<4.0",
    "anthropic>=0.40,<1.0",
    "langgraph>=0.2,<1.0",
    "langchain-anthropic>=0.3,<1.0",
    "langchain-core>=0.3,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.4,<1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apps*", "config*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
asyncio_mode = "auto"
